feat: adds tailscale service for secure networking
Adds a Tailscale service to the project, enabling secure
remote access when deployed to Balena.

This allows access to the intermediate server via a secure
Tailscale network, in addition to the local network.

Also updates the deployment process, and readme to document
how to configure tailscale during the balena deployment.

Updates the project structure, moving the node-server
related files into a dedicated directory.